% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{AnnenBatrute}
\alias{AnnenBatrute}
\title{AnnenBatrute}
\format{
\preformatted{Simple feature collection with 29 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -57709.25 ymin: 6735135 xmax: 1100141 ymax: 7940849
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 29 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64      100000 1997-06-30       (232372 7068014, 232180.5 7067841, 229664.3~
 2         64      100000 1997-06-30       (131289.9 7019729, 131034.7 7020006, 130766~
 3         64      100000 2009-06-18       (100137 6981371, 99885.69 6981211, 98024.83~
 4         64      100000 2009-06-18       (473607.3 7462664, 473568.4 7462671, 472290~
 5         64      100000 2009-06-18       (438552 7416365, 438181.6 7414906, 437773 7~
 6         64      100000 1997-06-30       (940512.8 7920810, 941631.6 7920313, 942925~
 7         64      100000 1997-06-30       (1043342 7908137, 1043203 7906704, 1042529 ~
 8         64      100000 1997-06-30       (815970.2 7862162, 816776.2 7862064, 818230~
 9         64      100000 1997-06-30       (815970.2 7862162, 814673.2 7862433, 814190~
10         64      100000 1997-06-30       (776929.4 7809142, 776228 7809268, 775527.1~
# ... with 19 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Samferdsel_GML.gml}
}
\usage{
AnnenBatrute
}
\description{
AnnenBåtrute
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
