% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{ApentOmrade}
\alias{ApentOmrade}
\title{ApentOmrade}
\format{
\preformatted{Simple feature collection with 756 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -76856.95 ymin: 6454300 xmax: 1113400 ymax: 7939874
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 756 x 2
   oppdateringsdato                                                            geometry
 * <date>                                                                 <POLYGON [m]>
 1 NA               ((298169.3 7089279, 297423.2 7088393, 294035 7086712, 290060.2 708~
 2 NA               ((277334.3 7064717, 278909.2 7063648, 276131.5 7061298, 273631.6 7~
 3 NA               ((238200 7059200, 238200 7058200, 233700 7057100, 230700 7059100, ~
 4 NA               ((290150.3 7057609, 290131.5 7057598, 287931.5 7056398, 285331.5 7~
 5 NA               ((155031.8 7040798, 149031.8 7041698, 147031.8 7044998, 140131.8 7~
 6 NA               ((242712.8 7053703, 243429.8 7051809, 244205.2 7050000, 243378 705~
 7 NA               ((234145.8 7045247, 228823.8 7043473, 226489.4 7046232, 228399.4 7~
 8 NA               ((216026.7 7047137, 217931.7 7047298, 222180.3 7047962, 225852.7 7~
 9 NA               ((242830.2 7037107, 238798 7034560, 235614.8 7036682, 236039.2 704~
10 NA               ((188931.1 7035565, 192300 7032900, 192539.2 7032900, 196600 70329~
# ... with 746 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
ApentOmrade
}
\description{
ÅpentOmråde
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
