% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Arealbrukgrense}
\alias{Arealbrukgrense}
\title{Arealbrukgrense}
\format{
\preformatted{Simple feature collection with 1226 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -52544.3 ymin: 6455812 xmax: 1113400 ymax: 7900065
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 1,226 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64      100000 1997-06-30       (1e+06 7899921, 1001977 7899415, 1002517 78~
 2         64      100000 1997-06-30                  (999786.4 7873646, 1e+06 7873671)
 3         64      100000 1997-06-30       (978407.4 7870327, 979990.7 7870723, 980300~
 4         64      100000 1997-06-30       (797810.6 7817676, 796300.9 7816730, 793429~
 5         64      100000 2005-09-15       (1050848 7815799, 1054240 7815071, 1055354 ~
 6         64      100000 1997-06-30       (1081008 7813116, 1083247 7814580, 1085904 ~
 7         64      100000 1997-06-30       (751282.4 7809793, 752100 7811800, 750700 7~
 8         64      100000 1997-06-30               (724198.4 7777000, 726529.6 7776386)
 9         64      100000 1997-06-30       (819788.8 7774214, 821328.9 7776048, 817310~
10         64      100000 1997-06-30       (8e+05 7771629, 799500 7771700, 796700 7772~
# ... with 1,216 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
Arealbrukgrense
}
\description{
Arealbrukgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
