% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Bane}
\alias{Bane}
\title{Bane}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 68 features and 4 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -42041.5 ymin: 6467475 xmax: 627357.9 ymax: 7594187
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 68 × 5
   medium oppdateringsdato målemetode nøyaktighet                            senterlinje
 * <chr>  <date>           <chr>            <int>                       <LINESTRING [m]>
 1 T      NA               gen             100000 (230221.1 6631260, 228223 6632547, 22…
 2 T      NA               gen             100000 (297139.9 7041486, 296140.5 7041425, …
 3 T      NA               gen             100000 (131782.1 6519434, 133081.2 6517759, …
 4 T      NA               gen             100000 (80689.58 6748733, 77993.17 6749641, …
 5 T      NA               gen             100000 (265459 6746662, 265233.8 6746434, 26…
 6 T      NA               gen             100000 (72361.6 6773286, 72749.63 6771242, 7…
 7 T      NA               gen             100000 (125410.4 6957524, 125552.8 6957398, …
 8 T      NA               gen             100000 (235206.8 6679379, 234738.9 6678992, …
 9 T      NA               gen             100000 (265133 6626840, 267774.6 6626155, 26…
10 T      NA               gen             100000 (131782.1 6519434, 130768.8 6519451, …
# ℹ 58 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Samferdsel_GML.gml}
}
\usage{
Bane
}
\description{
Bane
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
