% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Bilferjestrekning}
\alias{Bilferjestrekning}
\title{Bilferjestrekning}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 44 features and 6 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -48412.92 ymin: 6580811 xmax: 713984 ymax: 7735609
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 44 × 7
   målemetode nøyaktighet medium oppdateringsdato vegkategori vegnummer
 *      <int>       <int> <chr>  <date>           <chr>           <int>
 1         64      100000 T      NA               F                  17
 2         64      100000 T      NA               E                  39
 3         64      100000 T      NA               F                  NA
 4         64      100000 T      2022-03-01       F                  NA
 5         64      100000 T      2022-03-01       F                 549
 6         64      100000 T      2022-06-14       F                  NA
 7         64      100000 T      2022-03-01       R                 850
 8         64      100000 T      NA               F                  91
 9         64      100000 T      NA               F                  82
10         64      100000 T      NA               F                  17
# ℹ 34 more rows
# ℹ 1 more variable: senterlinje <LINESTRING [m]>
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Samferdsel_GML.gml}
}
\usage{
Bilferjestrekning
}
\description{
Bilferjestrekning
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
