% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Dataavgrensning}
\alias{Dataavgrensning}
\title{Dataavgrensning}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 265 features and 1 field
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -99438.22 ymin: 6426073 xmax: 1121927 ymax: 7962739
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 265 × 2
   oppdateringsdato                                                               grense
 * <date>                                                               <LINESTRING [m]>
 1 NA               (989380.5 7828123, 989400.3 7828072, 989723.1 7827239, 990585.6 782…
 2 NA                                               (993337.6 7818746, 993288.8 7818797)
 3 NA               (975433.7 7822196, 975430.2 7822192, 975106.3 7821845, 974730.4 781…
 4 NA                             (974641.1 7817485, 974670.2 7817495, 974710.7 7817899)
 5 NA               (974641.1 7817485, 974625.2 7817479, 972686 7816811, 972560.1 78167…
 6 NA                             (971135.8 7816277, 971166.2 7816287, 972560.1 7816767)
 7 NA                             (971135.8 7816277, 971119.6 7816271, 971101.2 7816238)
 8 NA               (967498.6 7812244, 967552.7 7812259, 969109.4 7812695, 971065.6 781…
 9 NA                               (966582.4 7811140, 966601.1 7811198, 966626 7811277)
10 NA                                (1109885 7812225, 1111178 7811050, 1111349 7811335)
# ℹ 255 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
Dataavgrensning
}
\description{
Dataavgrensning
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
