% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{FiktivDelelinje}
\alias{FiktivDelelinje}
\title{FiktivDelelinje}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 372 features and 1 field
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -94248.72 ymin: 6445086 xmax: 1114719 ymax: 7953516
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 372 × 2
   oppdateringsdato                                                               grense
 * <date>                                                               <LINESTRING [m]>
 1 NA                                                     (8e+05 7813454, 8e+05 7826565)
 2 NA                                                     (8e+05 7813392, 8e+05 7813454)
 3 NA                                                     (8e+05 7774430, 8e+05 7784300)
 4 NA                                                     (8e+05 7771629, 8e+05 7774430)
 5 NA                                                     (8e+05 7768450, 8e+05 7771629)
 6 NA               (715801.4 7800000, 727676.1 7800000, 730384.4 7800000, 734381.3 780…
 7 NA                                                     (8e+05 7800000, 8e+05 7813392)
 8 NA                                               (800901.6 7800000, 801701.7 7800000)
 9 NA                                                 (1056031 7800000, 1061280 7800000)
10 NA                                (1061280 7800000, 1062146 7800000, 1071917 7800000)
# ℹ 362 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
FiktivDelelinje
}
\description{
FiktivDelelinje
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
