% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Fylkesgrense}
\alias{Fylkesgrense}
\title{Fylkesgrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 178 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -86015.82 ymin: 6466105 xmax: 805691.2 ymax: 7853461
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 178 × 4
   målemetode nøyaktighet oppdateringsdato                                        grense
 *      <int>       <int> <date>                                        <LINESTRING [m]>
 1         64      100000 2023-12-18       (158596.2 7070657, 154320.8 7076843, 138033.…
 2         64      100000 2023-12-18                (708388.8 7840175, 701859.7 7853461)
 3         64      100000 2023-12-18                (168764.4 6766728, 170018.6 6762886)
 4         64      100000 2023-12-18       (525157.5 7603548, 525714.1 7602849, 526672.…
 5         64      100000 2023-12-18       (235189.7 6629201, 237056.3 6631240, 239182.…
 6         64      100000 2023-12-18                (281741.5 6953300, 284648.3 6956864)
 7         64      100000 2023-12-18       (213013.5 6615913, 214404.2 6617019, 216965.…
 8         64      100000 2023-12-18       (200583.7 6925816, 201141.7 6925605, 202629.…
 9         64      100000 2023-12-18       (784162.2 7725076, 788041.3 7726022, 790353.…
10         64      100000 2023-12-18                (613885.9 7612332, 623019.9 7613727)
# ℹ 168 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000AdministrativeOmrader_GML.gml}
}
\usage{
Fylkesgrense
}
\description{
Fylkesgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
