% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Havflate}
\alias{Havflate}
\title{Havflate}
\format{
\preformatted{Simple feature collection with 65 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -99438.22 ymin: 6426073 xmax: 1121927 ymax: 7962739
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 65 x 2
   oppdateringsdato                                                            geometry
 * <date>                                                                 <POLYGON [m]>
 1 NA               ((71732.01 6791199, 72832.02 6785999, 74932.02 6783999, 78032.02 6~
 2 NA               ((262231.7 6647700, 262531.7 6646000, 263331.7 6643000, 262808.9 6~
 3 NA                  ((1e+06 7876015, 1e+06 7879140, 1000341 7878997, 1e+06 7876015))
 4 NA               ((756766.8 7800000, 758956.1 7800000, 759930.1 7798296, 759230.1 7~
 5 NA               ((773579.2 7800000, 779238.4 7800000, 780930.1 7799696, 785530.1 7~
 6 NA               ((8e+05 7800000, 800901.6 7800000, 800359.4 7798753, 800030.1 7797~
 7 NA               ((1071917 7800000, 1074363 7800000, 1074736 7798611, 1076054 77937~
 8 NA                 ((6e+05 7710863, 600530.4 7710096, 6e+05 7710078, 6e+05 7710863))
 9 NA               ((6e+05 7651553, 6e+05 7654558, 601430.4 7654796, 602730.4 7652596~
10 NA               ((607856.8 7622010, 609330.4 7620796, 608530.4 7619796, 605630.4 7~
# ... with 55 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
Havflate
}
\description{
Havflate
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
