% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Innsjo}
\alias{Innsjo}
\title{Innsjo}
\format{
\preformatted{Simple feature collection with 49 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: 2432.2 ymin: 6494000 xmax: 1086878 ymax: 7792403
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 49 x 2
   oppdateringsdato                                                            geometry
 * <date>                                                                 <POLYGON [m]>
 1 NA               ((855481.2 7759655, 855630.4 7759596, 857430.4 7760096, 858930.5 7~
 2 NA               ((420282.8 7231164, 420814.7 7231046, 421487 7230898, 422131 72290~
 3 NA               ((432431 7207298, 431931 7204998, 430331 7205398, 429877.6 7207665~
 4 NA               ((157162.2 6915420, 157359.5 6916209, 157531.9 6916899, 160331.9 6~
 5 NA               ((52031.81 6896199, 48658.24 6896021, 48231.79 6895999, 45431.78 6~
 6 NA               ((305731.5 6814499, 306031.5 6811699, 305267.9 6812081, 304631.5 6~
 7 NA               ((144894.6 6799353, 146232 6799799, 146902.6 6799627, 150132 67987~
 8 NA               ((225515 6723752, 226397.2 6723676, 227678.9 6722309, 229898.3 672~
 9 NA               ((154832 6717099, 154232 6715899, 150632 6716199, 148532 6716899, ~
10 NA               ((91070.78 6648591, 93032.14 6648199, 97632.14 6645100, 99832.14 6~
# ... with 39 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
Innsjo
}
\description{
Innsjø
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
