% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Innsjokant}
\alias{Innsjokant}
\title{Innsjokant}
\format{
\preformatted{Simple feature collection with 170 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 2432.2 ymin: 6494000 xmax: 1085692 ymax: 7792403
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 170 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64      100000 1997-06-30       (859540.6 7756471, 860330.5 7756296, 860430~
 2         64      100000 1997-06-30       (687283.3 7608931, 689530.2 7608096, 695730~
 3         64      100000 1997-06-30       (860790.8 7749232, 860696.5 7749080, 860230~
 4         64      100000 1997-06-30       (661100.9 7622624, 661530.2 7623196, 663930~
 5         64      100000 1997-06-30       (663260.4 7619025, 662130.2 7619696, 660207~
 6         64      100000 1997-06-30       (420282.8 7231164, 420131.1 7231198, 420039~
 7         64      100000 1997-06-30       (458188.6 7306364, 458030.9 7306497, 460330~
 8         64      100000 1997-06-30       (446812 7304128, 448030.9 7303097, 449930.9~
 9         64      100000 1997-06-30       (458430.9 7283197, 455530.9 7283197, 454630~
10         64      100000 1997-06-30       (459609.8 7293091, 458030.9 7292097, 457130~
# ... with 160 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
Innsjokant
}
\description{
Innsjøkant
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
