% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kommune}
\alias{Kommune}
\title{Kommune}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 359 features and 3 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -99551.21 ymin: 6426048 xmax: 1121941 ymax: 7962744
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 359 × 4
   oppdateringsdato kommunenummer navn                                            område
 * <date>           <chr>         <chr>                                    <POLYGON [m]>
 1 2023-12-18       3110          Hvaler         ((265465.6 6562560, 261956.8 6562978, …
 2 2023-12-18       1563          Sunndal        ((165743 6994538, 164074.4 6996374, 16…
 3 2023-12-18       5046          Høylandet      ((386053.7 7206776, 385889.3 7209195, …
 4 2023-12-18       4612          Sveio          ((-29588.63 6658587, -28106.39 6661595…
 5 2023-12-18       3203          Asker          ((240950.6 6649596, 240763.5 6650012, …
 6 2023-12-18       4639          Vik            ((52184.7 6812682, 48633.17 6812808, 4…
 7 2023-12-18       1578          Fjord          ((102203.3 6950306, 100376.2 6952209, …
 8 2023-12-18       3316          Modum          ((217799.5 6681303, 217715.6 6681324, …
 9 2023-12-18       1865          Vågan          ((457447.7 7599335, 438606 7611106, 43…
10 2023-12-18       3226          Aurskog-Høland ((308071.8 6658838, 307821.5 6658879, …
# ℹ 349 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000AdministrativeOmrader_GML.gml}
}
\usage{
Kommune
}
\description{
Kommune
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
