% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kommunegrense}
\alias{Kommunegrense}
\title{Kommunegrense}
\format{
\preformatted{Simple feature collection with 867 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -92079.54 ymin: 6428945 xmax: 1111593 ymax: 7962288
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 867 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64      100000 2020-12-14       (407052.8 7146773, 388962 7145564, 384133.6~
 2         64      100000 2020-12-14       (407052.8 7146773, 409519 7137677, 413801.1~
 3         64      100000 2020-12-14       (63769.03 6760602, 63570.05 6759135, 63546.~
 4         64      100000 2020-12-14       (-16558.74 6692132, -16387.55 6691481, -163~
 5         64      100000 2020-12-14       (-16558.74 6692132, -16190.73 6693913, -159~
 6         64      100000 2020-12-14       (296205 7070123, 294556.3 7073785, 291040.9~
 7         64      100000 2020-12-14       (296205 7070123, 301369.7 7073079, 301649.8~
 8         64      100000 2020-12-14               (37936.65 6981824, 24150.37 6999320)
 9         64      100000 2020-12-14       (39544.67 6956626, 40560.26 6958958, 44152.~
10         64      100000 2020-12-14       (39544.67 6956626, 41856.4 6955434, 44267.8~
# ... with 857 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000AdministrativeOmrader_GML.gml}
}
\usage{
Kommunegrense
}
\description{
Kommunegrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
