% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kystkontur}
\alias{Kystkontur}
\title{Kystkontur}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 1240 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -76856.95 ymin: 6453900 xmax: 1104413 ymax: 7939874
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 1,240 × 4
   målemetode nøyaktighet oppdateringsdato                                        grense
 *      <int>       <int> <date>                                        <LINESTRING [m]>
 1         64      100000 NA                        (821701.6 7821959, 818533.5 7822839)
 2         64      100000 NA               (1011645 7841200, 1014342 7841698, 1016942 7…
 3         64      100000 NA               (998402.9 7898779, 996241 7900797, 995841.1 …
 4         64      100000 NA               (1097083 7886303, 1098151 7886897, 1097498 7…
 5         64      100000 NA                        (916132.7 7877796, 916402.2 7877894)
 6         64      100000 NA               (950002.3 7877773, 951435.1 7877396, 951068.…
 7         64      100000 NA               (916402.2 7877894, 922433.1 7880096, 928233.…
 8         64      100000 NA               (878232.5 7841555, 878031.1 7841496, 875931 …
 9         64      100000 NA               (8e+05 7813454, 800030.1 7813396, 8e+05 7813…
10         64      100000 NA               (814678.7 7813242, 815530.2 7813596, 817630.…
# ℹ 1,230 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
Kystkontur
}
\description{
Kystkontur
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
