% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Lufthavn}
\alias{Lufthavn}
\title{Lufthavn}
\format{
\preformatted{Simple feature collection with 14 features and 9 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -55217.49 ymin: 6472154 xmax: 651476.6 ymax: 7734999
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 14 x 10
   målemetode nøyaktighet oppdateringsdato iataKode icaoKode lufthavneier  lufthavntype
 *      <int>       <int> <date>           <chr>    <chr>    <chr>         <chr>       
 1         64      100000 2012-05-02       RYG      ENRY     Forsvarsdepa~ L           
 2         64      100000 2013-04-24       MOL      ENML     Avinor AS     L           
 3         64      100000 2012-05-02       HAU      ENHD     Avinor AS     L           
 4         64      100000 2012-05-02       BOO      ENBO     FLO Base Bod~ L           
 5         64      100000 2012-05-02       OLA      ENOL     FLO Base Ørl~ L           
 6         64      100000 2012-05-02       KRS      ENCN     Avinor AS     L           
 7         64      100000 2012-05-02       OSL      ENGM     Oslo Lufthav~ L           
 8         64      100000 2012-05-02       SVG      ENZV     Avinor AS     L           
 9         64      100000 2012-05-02       TRD      ENVA     Avinor AS     L           
10         64      100000 2012-05-02       BGO      ENBR     Avinor AS     L           
11         64      100000 2012-05-02       AES      ENAL     Avinor AS     L           
12         64      100000 2012-05-02       EVE      ENEV     Avinor AS     L           
13         64      100000 2012-05-02       TOS      ENTC     Avinor AS     L           
14         64      100000 2012-05-02       TRF      ENTO     Sandefjord l~ L           
# ... with 3 more variables: navn <chr>, trafikktype <chr>, geometry <POINT [m]>
}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
Lufthavn
}
\description{
Lufthavn
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
