% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N5000-package.R
\docType{package}
\name{N5000-package}
\alias{N5000}
\alias{N5000-package}
\title{N5000: Kartdata}
\source{
\itemize{
\item \url{https://kartkatalog.geonorge.no/metadata/kartverket/n5000-kartdata/c777d53d-8916-4d9d-bae4-6d5140e0c569}
}
}
\description{
Kartdata fra \href{https://kartverket.no/}{Kartverket}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hmalmedal/N5000}
  \item Report bugs at \url{https://github.com/hmalmedal/N5000/issues}
}

}
\author{
\strong{Maintainer}: Håkon Malmedal \email{hmalmedal@gmail.com}

Other contributors:
\itemize{
  \item Kartverket [copyright holder]
}

}
\keyword{package}
