% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Naturverngrense}
\alias{Naturverngrense}
\title{Naturverngrense}
\format{
\preformatted{Simple feature collection with 45 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 1183.26 ymin: 6472365 xmax: 1088054 ymax: 7893205
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 45 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64      100000 2016-02-01       (533613.6 7475526, 534740.7 7479364, 538105~
 2         64      100000 2016-02-01       (470434 7404044, 467625.7 7404774, 463616.3~
 3         64      100000 2016-02-01       (585497 7677287, 583700 7679367, 583253 768~
 4         64      100000 2016-02-01       (845960.9 7801580, 846043.4 7807370, 850022~
 5         64      100000 2016-02-01       (243785.7 6575539, 245050.8 6577008, 247536~
 6         64      100000 2016-02-01       (494981.1 7472050, 493041.1 7472155, 493042~
 7         64      100000 2016-02-01       (270385.5 6942384, 271351.6 6946091, 270891~
 8         64      100000 2016-02-01       (794168.3 7831570, 793987.3 7831691, 793915~
 9         64      100000 2016-02-01       (341250.5 6931621, 345746.7 6933494, 350427~
10         64      100000 2016-02-01       (99744.91 6881318, 97965.03 6874446, 100372~
# ... with 35 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Restriksjonsomrader_GML.gml}
}
\usage{
Naturverngrense
}
\description{
Naturverngrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
