% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Naturvernomrade}
\alias{Naturvernomrade}
\title{Naturvernomrade}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 42 features and 4 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: 1183.26 ymin: 6472365 xmax: 1088054 ymax: 7893205
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 42 × 5
   oppdateringsdato navn           vernedato  verneform                           område
 * <date>           <chr>          <date>     <chr>                        <POLYGON [m]>
 1 2017-01-04       Rago           1971-01-22 NP        ((533613.6 7475526, 549468.5 74…
 2 2017-01-04       Láhko          2012-12-14 NP        ((457971.3 7413120, 460540.9 74…
 3 2017-01-04       Ånderdalen     2004-06-04 NP        ((585497 7677287, 584761 767333…
 4 2017-01-04       Stabbursdalen  2002-12-20 NP        ((845960.9 7801580, 846028.5 77…
 5 2017-01-04       Færder         2013-08-23 NP        ((243785.7 6575539, 243406.8 65…
 6 2017-01-04       Sjunkhatten    2010-02-05 NP        ((494981.1 7472050, 495072.8 74…
 7 2017-01-04       Forollhogna    2001-12-21 NP        ((270385.5 6942384, 270875.7 69…
 8 2017-01-04       Seiland        2006-12-08 NP        ((794168.3 7831570, 794321 7831…
 9 2017-01-04       Femundsmarka   2003-02-21 NP        ((341250.5 6931621, 341922.7 69…
10 2017-01-04       Jostedalsbreen 1991-10-25 NP        ((87209.95 6861144, 84284.93 68…
# ℹ 32 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Restriksjonsomrader_GML.gml}
}
\usage{
Naturvernomrade
}
\description{
Naturvernområde
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
