% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Riksgrense}
\alias{Riksgrense}
\title{Riksgrense}
\format{
\preformatted{Simple feature collection with 36 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 295815.3 ymin: 6532318 xmax: 1115056 ymax: 7828443
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 36 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64      100000 2020-12-14       (1008977 7804953, 1011539 7802961, 1012505 ~
 2         64      100000 2020-12-14                 (1008765 7808532, 1008977 7804953)
 3         64      100000 2020-12-14       (931415.7 7785646, 931769.7 7785963, 932947~
 4         64      100000 2020-12-14       (916230.4 7677406, 916837.2 7677804, 920651~
 5         64      100000 2020-12-14       (777865.5 7678331, 786334 7667172, 786999.9~
 6         64      100000 2020-12-14       (747182.8 7701826, 750698 7701151, 753776.2~
 7         64      100000 2020-12-14       (738347 7692234, 740578.4 7695439, 747182.8~
 8         64      100000 2020-12-14       (701756.5 7667791, 705043.6 7668334, 711360~
 9         64      100000 2020-12-14       (701248.5 7613149, 705613.4 7617213, 709539~
10         64      100000 2020-12-14       (296708.5 6544230, 296654.2 6542434, 296433~
# ... with 26 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000AdministrativeOmrader_GML.gml}
}
\usage{
Riksgrense
}
\description{
Riksgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
