% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Skog}
\alias{Skog}
\title{Skog}
\format{
\preformatted{Simple feature collection with 429 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -52544.3 ymin: 6453900 xmax: 1114945 ymax: 7900065
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 429 x 2
   oppdateringsdato                                                            geometry
 * <date>                                                                 <POLYGON [m]>
 1 2011-03-24       ((140500 6999700, 139200 6996900, 141800 6993300, 143000 6993900, ~
 2 2011-03-24       ((328938.2 6731425, 332365.2 6728835, 332396.9 6728866, 333202.1 6~
 3 2011-03-24       ((1027985 7805159, 1028492 7803550, 1029352 7800706, 1029164 77982~
 4 NA               ((561682 7541539, 562530.6 7540397, 563830.6 7540797, 563630.6 754~
 5 NA               ((499224.6 7518328, 498030.8 7517997, 492430.8 7517597, 493230.8 7~
 6 NA               ((515230.7 7511197, 519730.7 7508497, 521855.3 7509094, 522173.1 7~
 7 NA               ((460664 7439736, 457155.1 7441616, 456630.9 7441897, 457230.9 744~
 8 NA               ((430008.2 7414879, 431504 7415254, 434013.1 7415254, 435072 74152~
 9 NA               ((429617.7 7404671, 428176.2 7406768, 429432.4 7408373, 429583 740~
10 NA               ((429570.9 7400000, 424395.5 7400000, 423831 7400897, 422331 74017~
# ... with 419 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
Skog
}
\description{
Skog
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
