% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{SnoIsbre}
\alias{SnoIsbre}
\title{SnoIsbre}
\format{
\preformatted{Simple feature collection with 25 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: 2067.32 ymin: 6675296 xmax: 771816.9 ymax: 7804397
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 25 x 2
   oppdateringsdato                                                            geometry
 * <date>                                                                 <POLYGON [m]>
 1 NA               ((483173.8 7418825, 483258 7418789, 480703 7414786, 478738 7411160~
 2 NA               ((580374.8 7541221, 579062.3 7541527, 576828 7542164, 575133 75423~
 3 NA               ((159841.7 6844159, 161222.7 6839853, 160850.9 6837785, 159687.3 6~
 4 NA               ((28330.55 6695316, 26947.29 6694737, 26661.58 6695300, 25699.64 6~
 5 2009-05-12       ((134635.5 6839884, 134205.4 6839737, 132567.9 6841359, 130306.3 6~
 6 2009-05-12       ((124574.1 6836164, 124744.3 6834271, 123903.1 6832456, 123833 683~
 7 2009-05-12       ((141194.4 6867630, 141734 6866281, 138908 6863408, 134532.7 68644~
 8 2009-05-12       ((472675.7 7321332, 473170.5 7319243, 472329.9 7317751, 471823.6 7~
 9 2009-05-12       ((593836 7566609, 592086 7565516, 591881.5 7565553, 588148 7566227~
10 2009-05-12       ((560488.3 7444281, 560470.1 7443340, 555221.3 7444163, 554496.7 7~
# ... with 15 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
SnoIsbre
}
\description{
SnøIsbre
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
