% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Stasjon}
\alias{Stasjon}
\title{Stasjon}
\format{
\preformatted{Simple feature collection with 56 features and 4 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -40191.86 ymin: 6467651 xmax: 625944.6 ymax: 7594187
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 56 x 5
   målemetode nøyaktighet oppdateringsdato navn                               geometry
 *      <int>       <int> <date>           <chr>                           <POINT [m]>
 1         64      100000 2004-11-30       Oslo Lufthavn Gardermoen (283584.3 6679144)
 2         64      100000 2009-05-12       Kongsvinger              (334071.1 6675162)
 3         64      100000 2009-05-12       Tønsberg                 (238512.8 6579249)
 4         64      100000 2009-05-12       Larvik                   (215110.7 6556333)
 5         64      100000 2004-11-30       Koppang                  (290000.8 6833010)
 6         64      100000 2012-05-07       Hamar                    (286491.4 6746722)
 7         64      100000 2009-05-12       Dombås                   (193736.8 6895785)
 8         64      100000 2009-05-12       Kristiansand             (86015.71 6467651)
 9         64      100000 2009-05-12       Tynset                   (280967.8 6911296)
10         64      100000 2004-11-30       Bjørnfjell               (625944.6 7594187)
# ... with 46 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Samferdsel_GML.gml}
}
\usage{
Stasjon
}
\description{
Stasjon
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
