% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{StedsnavnTekst}
\alias{StedsnavnTekst}
\title{StedsnavnTekst}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 165 features and 17 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -55504.6 ymin: 6443620 xmax: 1101454 ymax: 7939718
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 165 × 18
   streng  oppdateringsdato datafangstdato tekstReferansePunktN…¹ tekstReferansePunktØst
 * <chr>   <date>           <date>                          <int>                  <int>
 1 Tromsø  2021-11-11       2017-01-04                          0                      0
 2 Hammer… 2021-11-11       2017-01-04                          0                      0
 3 Honnin… 2021-11-11       2017-01-04                          0                      0
 4 Vardø   2021-11-11       2017-01-04                          0                      0
 5 Kirken… 2021-11-11       2017-01-04                          0                      0
 6 Alta    2021-11-11       2017-01-04                          0                      0
 7 Brønnø… 2021-11-11       2017-01-04                          0                      2
 8 Narvik  2021-11-11       2017-01-04                          0                      0
 9 Svolvær 2021-11-11       2017-01-04                          0                      0
10 Mo i R… 2021-11-11       2017-01-04                          0                      0
# ℹ 155 more rows
# ℹ abbreviated name: ¹​tekstReferansePunktNord
# ℹ 13 more variables: tegnavstand <int>, skriftkode <int>, skrifttype <chr>,
#   referansemålestokk <int>, fulltekst <chr>, navneobjekthovedgruppe <int>,
#   navneobjektgruppe <int>, navneobjekttype <int>, språk <chr>,
#   stedsnavnsnummer <int>, skrivemåtenummer <int>, stedsnummer <int>,
#   geometri <LINESTRING [m]>
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Stedsnavn_GML.gml}
}
\usage{
StedsnavnTekst
}
\description{
StedsnavnTekst
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
