% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Territorialgrense}
\alias{Territorialgrense}
\title{Territorialgrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 78 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -99551.21 ymin: 6426048 xmax: 1121941 ymax: 7962744
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 78 × 4
   målemetode nøyaktighet oppdateringsdato                                        grense
 *      <int>       <int> <date>                                        <LINESTRING [m]>
 1         64      100000 2023-12-18       (-86015.82 6648235, -80327.37 6659428, -7969…
 2         64      100000 2023-12-18                (362047.3 7414276, 354846.2 7437006)
 3         64      100000 2023-12-18       (394959.1 7562014, 400056.4 7571103, 400564.…
 4         64      100000 2023-12-18       (-86165.9 6792747, -87860.58 6809231, -87940…
 5         64      100000 2023-12-18       (764795.7 7891994, 770308.5 7896821, 770685.…
 6         64      100000 2023-12-18               (-25091.8 6958283, -20575.89 6963321)
 7         64      100000 2023-12-18       (500870.5 7689595, 527306.2 7710050, 528110.…
 8         64      100000 2023-12-18       (253149.8 7193866, 263118.3 7203317, 263404.…
 9         64      100000 2023-12-18       (548801.2 7721400, 565026.7 7729633, 579406.…
10         64      100000 2023-12-18       (-36166.04 6482741, -48924.89 6497525, -4956…
# ℹ 68 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000AdministrativeOmrader_GML.gml}
}
\usage{
Territorialgrense
}
\description{
Territorialgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
