% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Tettbebyggelse}
\alias{Tettbebyggelse}
\title{Tettbebyggelse}
\format{
\preformatted{Simple feature collection with 125 features and 4 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -51717 ymin: 6457412 xmax: 1097078 ymax: 7930522
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 125 x 5
   målemetode nøyaktighet oppdateringsdato tettbebyggelsestype             geometry
 *      <int>       <int> <date>           <chr>                        <POINT [m]>
 1         64      100000 2017-09-27       MindreTettbebyggelse    (-51717 6613113)
 2         64      100000 2017-09-27       MindreTettbebyggelse    (-13502 6577473)
 3         64      100000 2017-09-27       MindreTettbebyggelse  (15740.38 6644117)
 4         64      100000 2017-09-27       MindreTettbebyggelse  (73774.92 6586314)
 5         64      100000 2017-09-27       MindreTettbebyggelse     (82682 6516692)
 6         64      100000 2017-09-27       MindreTettbebyggelse (-41630.48 6548546)
 7         64      100000 2017-09-27       MindreTettbebyggelse    (-24003 6515513)
 8         64      100000 2017-09-27       MindreTettbebyggelse     (13327 6491769)
 9         64      100000 2017-09-27       MindreTettbebyggelse  (16332.21 6468497)
10         64      100000 2017-09-27       MindreTettbebyggelse  (31951.71 6469561)
# ... with 115 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
Tettbebyggelse
}
\description{
Tettbebyggelse
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
