% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Veglenke}
\alias{Veglenke}
\title{Veglenke}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 715 features and 9 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -58762.81 ymin: 6457040 xmax: 1097078 ymax: 7930871
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 715 × 10
   oppdateringsdato målemetode nøyaktighet typeVeg  vegkategori vegnummer medium vegfase
 * <date>           <chr>            <int> <chr>    <chr>           <int> <chr>  <chr>  
 1 NA               gen             100000 bilferje F                  17 <NA>   <NA>   
 2 NA               gen             100000 bilferje E                  39 <NA>   <NA>   
 3 NA               gen             100000 bilferje F                  NA <NA>   <NA>   
 4 2022-03-01       gen             100000 bilferje F                  NA <NA>   <NA>   
 5 2022-03-01       gen             100000 bilferje F                 549 <NA>   <NA>   
 6 2022-06-14       gen             100000 bilferje F                  NA <NA>   <NA>   
 7 2022-03-01       gen             100000 bilferje R                 850 <NA>   <NA>   
 8 NA               gen             100000 bilferje F                  91 <NA>   <NA>   
 9 NA               gen             100000 bilferje F                  82 <NA>   <NA>   
10 NA               gen             100000 bilferje F                  17 <NA>   <NA>   
# ℹ 705 more rows
# ℹ 2 more variables: motorvegtype <chr>, senterlinje <LINESTRING [m]>
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Samferdsel_GML.gml}
}
\usage{
Veglenke
}
\description{
Veglenke
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
