% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{ApentOmrade}
\alias{ApentOmrade}
\title{ApentOmrade}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 756 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -76856.95 ymin: 6454300 xmax: 1113400 ymax: 7939874
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 756 × 2
   oppdateringsdato                                                               område
 * <date>                                                                  <POLYGON [m]>
 1 NA               ((298169.3 7089279, 294431.5 7087698, 290931.5 7086498, 288189.8 70…
 2 NA               ((277334.3 7064717, 276000 7067100, 273500 7067300, 271000 7068700,…
 3 NA               ((238200 7059200, 234300 7059400, 230800 7062200, 231800 7063100, 2…
 4 NA               ((290150.3 7057609, 289304.9 7058183, 288649.5 7058628, 286623.8 70…
 5 NA               ((155031.8 7040798, 160231.8 7044698, 157231.8 7056698, 145231.8 70…
 6 NA               ((242712.8 7053703, 239887.7 7051081, 243378 7050208, 244205.2 7050…
 7 NA               ((234145.8 7045247, 234553.7 7045383, 233492.6 7049203, 230573.2 70…
 8 NA               ((216026.7 7047137, 216726.7 7045737, 215029.6 7045171, 215241.9 70…
 9 NA               ((242830.2 7037107, 242405.7 7039229, 241684.5 7040130, 240708 7041…
10 NA               ((188931.1 7035565, 184601.5 7035635, 184122.5 7036810, 179740.1 70…
# ℹ 746 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
ApentOmrade
}
\description{
ÅpentOmråde
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
