% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Arealbrukgrense}
\alias{Arealbrukgrense}
\title{Arealbrukgrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 1226 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -52544.3 ymin: 6455812 xmax: 1113400 ymax: 7900065
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 1,226 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 NA               gen             100000 (1e+06 7899921, 1001977 7899415, 1002517 789…
 2 NA               gen             100000             (999786.4 7873646, 1e+06 7873671)
 3 NA               gen             100000 (978407.4 7870327, 979990.7 7870723, 980300 …
 4 NA               gen             100000 (797810.6 7817676, 796300.9 7816730, 793429.…
 5 NA               gen             100000 (1050848 7815799, 1054240 7815071, 1055354 7…
 6 NA               gen             100000 (1081008 7813116, 1083247 7814580, 1085904 7…
 7 NA               gen             100000 (751282.4 7809793, 752100 7811800, 750700 78…
 8 NA               gen             100000          (724198.4 7777000, 726529.6 7776386)
 9 NA               gen             100000 (819788.8 7774214, 821328.9 7776048, 817310.…
10 NA               gen             100000 (8e+05 7771629, 799500 7771700, 796700 77728…
# ℹ 1,216 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
Arealbrukgrense
}
\description{
Arealbrukgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
