% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Bilferjestrekning}
\alias{Bilferjestrekning}
\title{Bilferjestrekning}
\format{
\preformatted{Simple feature collection with 45 features and 6 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -48412.92 ymin: 6574056 xmax: 713984 ymax: 7735609
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 45 x 7
   målemetode nøyaktighet medium oppdateringsdato vegkategori vegnummer
 *      <int>       <int> <chr>  <date>           <chr>           <int>
 1         64      100000 T      2010-07-06       F                  17
 2         64      100000 T      2009-06-18       E                  39
 3         64      100000 T      2010-07-06       F                  91
 4         64      100000 T      2010-07-06       F                  86
 5         64      100000 T      2004-01-23       R                  83
 6         64      100000 T      2010-07-06       F                  82
 7         64      100000 T      2010-07-06       F                  17
 8         64      100000 T      2010-07-06       F                  17
 9         64      100000 T      2010-07-06       F                  17
10         64      100000 T      2013-04-24       R                  80
# ... with 35 more rows, and 1 more variable: geometry <LINESTRING [m]>
}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Samferdsel_GML.gml}
}
\usage{
Bilferjestrekning
}
\description{
Bilferjestrekning
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
