% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{ElvBekk}
\alias{ElvBekk}
\title{ElvBekk}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 137 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -4867.83 ymin: 6456207 xmax: 1085502 ymax: 7841997
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 137 × 4
   oppdateringsdato målemetode nøyaktighet                                   senterlinje
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 NA               gen             100000 (930915.8 7784519, 930455.1 7783030, 930356.…
 2 NA               gen             100000 (679230.2 7663696, 677730.2 7663396, 677395.…
 3 NA               gen             100000 (305190.7 6784040, 305231.5 6783899, 304689.…
 4 NA               gen             100000 (369131.2 7151598, 368421.5 7151449, 364831.…
 5 NA               gen             100000 (102932 6871299, 100832 6868499, 97235.66 68…
 6 NA               gen             100000 (318382.5 6954362, 317631.4 6953298, 316531.…
 7 NA               gen             100000 (828230.1 7659996, 828106.6 7660025, 826607.…
 8 NA               gen             100000 (309709.5 7043102, 309831.4 7043098, 310376 …
 9 NA               gen             100000 (309709.5 7043102, 310431.4 7041898, 311201.…
10 NA               gen             100000 (216113 6954031, 216231.8 6954098, 216517.1 …
# ℹ 127 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
ElvBekk
}
\description{
ElvBekk
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
