% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Fylkesgrense}
\alias{Fylkesgrense}
\title{Fylkesgrense}
\format{
\preformatted{Simple feature collection with 158 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -86015.82 ymin: 6466105 xmax: 627600.6 ymax: 7721400
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 158 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64      100000 2020-12-14               (295281.7 7268048, 315239.5 7258254)
 2         64      100000 2020-12-14       (156661.5 6772647, 155393.7 6774928, 155584~
 3         64      100000 2020-12-14       (168764.4 6766728, 168616.1 6766747, 167749~
 4         64      100000 2020-12-14               (170018.6 6762886, 168764.4 6766728)
 5         64      100000 2020-12-14       (187434.1 6750573, 184523.5 6752287, 178745~
 6         64      100000 2020-12-14       (190047.6 6742680, 190726.5 6743667, 190837~
 7         64      100000 2020-12-14       (193884.4 6728129, 194721.1 6730176, 194039~
 8         64      100000 2020-12-14       (215836.7 6713407, 214953.3 6713340, 214192~
 9         64      100000 2020-12-14       (228557.9 6732064, 226717.4 6732522, 223814~
10         64      100000 2020-12-14       (232646.5 6720966, 232551.3 6721160, 231556~
# ... with 148 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000AdministrativeOmrader_GML.gml}
}
\usage{
Fylkesgrense
}
\description{
Fylkesgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
