% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Fylkesgrense}
\alias{Fylkesgrense}
\title{Fylkesgrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 178 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -86015.82 ymin: 6466105 xmax: 805691.2 ymax: 7853461
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 178 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 2023-12-18       gen             100000          (595331.1 7607386, 603726.5 7611575)
 2 2023-12-18       gen             100000          (708388.8 7840175, 701859.7 7853461)
 3 2023-12-18       gen             100000          (613885.9 7612332, 623019.9 7613727)
 4 2023-12-18       gen             100000 (148486.5 6929627, 145658 6926498, 143384.6 …
 5 2023-12-18       gen             100000 (392079.3 7205123, 392674 7204621, 393588 72…
 6 2023-12-18       gen             100000 (230031.8 6917129, 230086 6917117, 230290 69…
 7 2023-12-18       gen             100000 (274413.9 6625377, 273952.1 6625459, 273615.…
 8 2023-12-18       gen             100000          (281741.5 6953300, 284648.3 6956864)
 9 2023-12-18       gen             100000 (301459.1 6953844, 302019.2 6952422, 303345.…
10 2023-12-18       gen             100000          (168764.4 6766728, 170018.6 6762886)
# ℹ 168 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000AdministrativeOmrader_GML.gml}
}
\usage{
Fylkesgrense
}
\description{
Fylkesgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
