% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Havflate}
\alias{Havflate}
\title{Havflate}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 65 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -99438.22 ymin: 6426073 xmax: 1121927 ymax: 7962739
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 65 × 2
   oppdateringsdato                                                               område
 * <date>                                                                  <POLYGON [m]>
 1 NA               ((71732.01 6791199, 71327.46 6793279, 71032 6794799, 71321.55 67956…
 2 NA               ((262231.7 6647700, 259619.7 6648925, 258253.7 6649519, 257831.7 66…
 3 NA                   ((1e+06 7876015, 1000341 7878997, 1e+06 7879140, 1e+06 7876015))
 4 NA               ((756766.8 7800000, 757430.1 7798596, 759230.1 7795696, 759930.1 77…
 5 NA               ((773579.2 7800000, 773830.1 7798696, 771430.1 7797196, 773230.1 77…
 6 NA               ((8e+05 7800000, 8e+05 7797973, 800030.1 7797996, 800359.4 7798753,…
 7 NA               ((1071917 7800000, 1072753 7796300, 1074653 7792100, 1076054 779370…
 8 NA                  ((6e+05 7710863, 6e+05 7710078, 600530.4 7710096, 6e+05 7710863))
 9 NA               ((6e+05 7651553, 601030.4 7651896, 602730.4 7652596, 601430.4 76547…
10 NA               ((607856.8 7622010, 607630.4 7622196, 603930.4 7623196, 600930.4 76…
# ℹ 55 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
Havflate
}
\description{
Havflate
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
