% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Innsjo}
\alias{Innsjo}
\title{Innsjo}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 49 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: 2432.2 ymin: 6494000 xmax: 1086878 ymax: 7792403
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 49 × 2
   oppdateringsdato                                                               område
 * <date>                                                                  <POLYGON [m]>
 1 NA               ((855481.2 7759655, 853630.4 7760396, 851930.4 7759796, 850730.4 77…
 2 NA               ((420282.8 7231164, 420131.1 7231198, 420039.2 7231006, 419031.1 72…
 3 NA               ((432431 7207298, 433731 7207998, 435578 7208386, 438531 7208398, 4…
 4 NA               ((157162.2 6915420, 157131.9 6915299, 159031.9 6914299, 160731.9 69…
 5 NA               ((52031.81 6896199, 53169.15 6896980, 53631.82 6897299, 54131.82 68…
 6 NA               ((305731.5 6814499, 305443 6815233, 304631.5 6817299, 303831.5 6819…
 7 NA               ((144894.6 6799353, 144732 6799299, 147332 6797499, 151232 6795499,…
 8 NA               ((225515 6723752, 225240.6 6723776, 224521.6 6723309, 224649.5 6723…
 9 NA               ((154832 6717099, 154686.5 6717199, 153232 6718199, 150232 6718699,…
10 NA               ((91070.78 6648591, 91032.14 6648599, 91062.17 6648559, 93132.14 66…
# ℹ 39 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
Innsjo
}
\description{
Innsjø
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
