% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Innsjokant}
\alias{Innsjokant}
\title{Innsjokant}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 170 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 2432.2 ymin: 6494000 xmax: 1085692 ymax: 7792403
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 170 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 NA               gen             100000 (859540.6 7756471, 860330.5 7756296, 860430.…
 2 NA               gen             100000 (687283.3 7608931, 689530.2 7608096, 695730.…
 3 NA               gen             100000 (860790.8 7749232, 860696.5 7749080, 860230.…
 4 NA               gen             100000 (661100.9 7622624, 661530.2 7623196, 663930.…
 5 NA               gen             100000 (663260.4 7619025, 662130.2 7619696, 660207.…
 6 NA               gen             100000 (459609.8 7293091, 458030.9 7292097, 457130.…
 7 NA               gen             100000 (420282.8 7231164, 420131.1 7231198, 420039.…
 8 NA               gen             100000 (458188.6 7306364, 458030.9 7306497, 460330.…
 9 NA               gen             100000 (446812 7304128, 448030.9 7303097, 449930.9 …
10 NA               gen             100000 (458430.9 7283197, 455530.9 7283197, 454630.…
# ℹ 160 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
Innsjokant
}
\description{
Innsjøkant
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
