% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kommune}
\alias{Kommune}
\title{Kommune}
\format{
\preformatted{Simple feature collection with 357 features and 3 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -99551.21 ymin: 6426048 xmax: 1121941 ymax: 7962744
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 357 x 4
   oppdateringsdato kommunenummer navn                                         geometry
 * <date>           <chr>         <chr>                                   <POLYGON [m]>
 1 2020-12-14       1506          Molde    ((115659.8 6989493, 119041.8 6989434, 11913~
 2 2020-12-14       5022          Rennebu  ((244139.3 6992547, 246010.1 6991573, 24604~
 3 2020-12-14       1563          Sunndal  ((165743 6994538, 168376.7 6990439, 171205.~
 4 2020-12-14       5026          Holtålen ((302803.8 6998076, 308424.7 6992390, 31126~
 5 2020-12-14       1532          Giske    ((37936.65 6981824, 47988.54 6970227, 46231~
 6 2020-12-14       1557          Gjemnes  ((139134.7 7004594, 140612.8 7002655, 14517~
 7 2020-12-14       5027          Midtre ~ ((285798.5 7007406, 290415.3 7003734, 29369~
 8 2020-12-14       5029          Skaun    ((253271 7034973, 254621.9 7033037, 257984.~
 9 2020-12-14       5033          Tydal    ((351680 7009423, 355662.1 6998417, 359130.~
10 2020-12-14       1566          Surnadal ((196731.6 7013732, 196827.5 7013311, 19683~
# ... with 347 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000AdministrativeOmrader_GML.gml}
}
\usage{
Kommune
}
\description{
Kommune
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
