% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kommune}
\alias{Kommune}
\title{Kommune}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 359 features and 5 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -99551.21 ymin: 6426048 xmax: 1121941 ymax: 7962744
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 359 × 6
   oppdateringsdato kommunenummer kommunenavn fylkesnummer fylkesnavn     
 * <date>           <chr>         <chr>       <chr>        <chr>          
 1 2023-12-18       1106          Haugesund   11           Rogaland       
 2 2023-12-18       4632          Austrheim   46           Vestland       
 3 2023-12-18       4018          Nome        40           Telemark       
 4 2023-12-18       4227          Kvinesdal   42           Agder          
 5 2023-12-18       1133          Hjelmeland  11           Rogaland       
 6 2023-12-18       1508          Ålesund     15           Møre og Romsdal
 7 2023-12-18       4220          Bygland     42           Agder          
 8 2023-12-18       4612          Sveio       46           Vestland       
 9 2023-12-18       4614          Stord       46           Vestland       
10 2023-12-18       5546          Kvænangen   55           Troms          
# ℹ 349 more rows
# ℹ 1 more variable: område <POLYGON [m]>
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000AdministrativeOmrader_GML.gml}
}
\usage{
Kommune
}
\description{
Kommune
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
