% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kommunegrense}
\alias{Kommunegrense}
\title{Kommunegrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 779 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -92079.54 ymin: 6428945 xmax: 1111593 ymax: 7962288
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 779 × 4
   målemetode nøyaktighet oppdateringsdato                                        grense
 *      <int>       <int> <date>                                        <LINESTRING [m]>
 1         64      100000 2023-12-18       (-2827.95 6577269, -2671.8 6573522, -4840.33…
 2         64      100000 2023-12-18       (70496.66 6795696, 70004.72 6792643, 68186.9…
 3         64      100000 2023-12-18       (-36166.04 6482741, -19906.89 6502689, -1548…
 4         64      100000 2023-12-18       (148438.3 6615423, 150634.5 6615890, 152168.…
 5         64      100000 2023-12-18       (402160 7463870, 396030 7459930, 384820 7447…
 6         64      100000 2023-12-18       (156661.5 6772647, 157490.7 6774025, 159214.…
 7         64      100000 2023-12-18       (324141.3 6709900, 324057.2 6709357, 323863 …
 8         64      100000 2023-12-18       (6831.96 6527943, 5276.47 6526828, 4261.85 6…
 9         64      100000 2023-12-18       (328001.5 6898424, 332356.6 6897002, 332481.…
10         64      100000 2023-12-18       (503140 7500160, 507750 7495180, 508000 7492…
# ℹ 769 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000AdministrativeOmrader_GML.gml}
}
\usage{
Kommunegrense
}
\description{
Kommunegrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
