% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kommunegrense}
\alias{Kommunegrense}
\title{Kommunegrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 779 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -92079.54 ymin: 6428945 xmax: 1111593 ymax: 7962288
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 779 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 2023-12-18       gen             100000          (315989.7 6738700, 315212.4 6731389)
 2 2023-12-18       gen             100000 (154256.5 6531598, 154418.6 6530770, 156200.…
 3 2023-12-18       gen             100000 (337699.8 7055861, 342126.2 7060072, 350078 …
 4 2023-12-18       gen             100000            (316288 6720699, 315212.4 6731389)
 5 2023-12-18       gen             100000 (215911.5 6680633, 215901.8 6680085, 215585 …
 6 2023-12-18       gen             100000        (-34198.39 6554681, -30076.67 6551892)
 7 2023-12-18       gen             100000          (40057.16 6588165, 43414.25 6588622)
 8 2023-12-18       gen             100000 (419050 7353890, 421300 7354000, 423569 7353…
 9 2023-12-18       gen             100000 (39544.67 6956626, 40560.26 6958958, 46059.4…
10 2023-12-18       gen             100000 (177520.6 6706275, 177228.6 6705839, 176817.…
# ℹ 769 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000AdministrativeOmrader_GML.gml}
}
\usage{
Kommunegrense
}
\description{
Kommunegrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
