% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kystkontur}
\alias{Kystkontur}
\title{Kystkontur}
\format{
\preformatted{Simple feature collection with 1240 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -76856.95 ymin: 6453900 xmax: 1104413 ymax: 7939874
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 1,240 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64      100000 1997-06-30               (821701.6 7821959, 818533.5 7822839)
 2         64      100000 1997-06-30       (1011645 7841200, 1014342 7841698, 1016942 ~
 3         64      100000 1997-06-30       (998402.9 7898779, 996241 7900797, 995841.1~
 4         64      100000 1997-06-30       (1097083 7886303, 1098151 7886897, 1097498 ~
 5         64      100000 1997-06-30               (916132.7 7877796, 916402.2 7877894)
 6         64      100000 1997-06-30       (950002.3 7877773, 951435.1 7877396, 951068~
 7         64      100000 1997-06-30       (916402.2 7877894, 922433.1 7880096, 928233~
 8         64      100000 1997-06-30       (878232.5 7841555, 878031.1 7841496, 875931~
 9         64      100000 1997-06-30       (8e+05 7813454, 800030.1 7813396, 8e+05 781~
10         64      100000 1997-06-30       (814678.7 7813242, 815530.2 7813596, 817630~
# ... with 1,230 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
Kystkontur
}
\description{
Kystkontur
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
