% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Lufthavn}
\alias{Lufthavn}
\title{Lufthavn}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 14 features and 9 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -55217.49 ymin: 6472154 xmax: 651476.6 ymax: 7734999
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 14 × 10
   målemetode nøyaktighet oppdateringsdato iataKode icaoKode lufthavneier   lufthavntype
 *      <int>       <int> <date>           <chr>    <chr>    <chr>          <chr>       
 1         64      100000 NA               RYG      ENRY     Forsvarsdepar… L           
 2         64      100000 NA               MOL      ENML     Avinor AS      L           
 3         64      100000 NA               HAU      ENHD     Avinor AS      L           
 4         64      100000 NA               BOO      ENBO     FLO Base Bodø… L           
 5         64      100000 NA               OLA      ENOL     FLO Base Ørla… L           
 6         64      100000 NA               KRS      ENCN     Avinor AS      L           
 7         64      100000 NA               OSL      ENGM     Oslo Lufthavn… L           
 8         64      100000 NA               SVG      ENZV     Avinor AS      L           
 9         64      100000 NA               TRD      ENVA     Avinor AS      L           
10         64      100000 NA               BGO      ENBR     Avinor AS      L           
11         64      100000 NA               AES      ENAL     Avinor AS      L           
12         64      100000 NA               EVE      ENEV     Avinor AS      L           
13         64      100000 NA               TOS      ENTC     Avinor AS      L           
14         64      100000 NA               TRF      ENTO     Sandefjord lu… L           
# ℹ 3 more variables: navn <chr>, trafikktype <chr>, posisjon <POINT [m]>
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
Lufthavn
}
\description{
Lufthavn
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
