% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Lufthavn}
\alias{Lufthavn}
\title{Lufthavn}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 14 features and 9 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -55217.49 ymin: 6472154 xmax: 651476.6 ymax: 7734999
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 14 × 10
   oppdateringsdato målemetode nøyaktighet lufthavntype trafikktype iataKode icaoKode
 * <date>           <chr>            <int> <chr>        <chr>       <chr>    <chr>   
 1 NA               gen             100000 L            I           RYG      ENRY    
 2 NA               gen             100000 L            I           MOL      ENML    
 3 NA               gen             100000 L            I           HAU      ENHD    
 4 NA               gen             100000 L            I           BOO      ENBO    
 5 NA               gen             100000 L            I           OLA      ENOL    
 6 NA               gen             100000 L            I           KRS      ENCN    
 7 NA               gen             100000 L            I           OSL      ENGM    
 8 NA               gen             100000 L            I           SVG      ENZV    
 9 NA               gen             100000 L            I           TRD      ENVA    
10 NA               gen             100000 L            I           BGO      ENBR    
11 NA               gen             100000 L            I           AES      ENAL    
12 NA               gen             100000 L            I           EVE      ENEV    
13 NA               gen             100000 L            I           TOS      ENTC    
14 NA               gen             100000 L            I           TRF      ENTO    
# ℹ 3 more variables: lufthavneier <chr>, navn <chr>, posisjon <POINT [m]>
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
Lufthavn
}
\description{
Lufthavn
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
