% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Naturverngrense}
\alias{Naturverngrense}
\title{Naturverngrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 45 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 1183.26 ymin: 6472365 xmax: 1088054 ymax: 7893205
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 45 × 4
   målemetode nøyaktighet oppdateringsdato                                        grense
 *      <int>       <int> <date>                                        <LINESTRING [m]>
 1         64      100000 NA               (863873 7658542, 863904.8 7658597, 864120.4 …
 2         64      100000 NA               (410167.3 7531961, 411245.7 7531237, 412145.…
 3         64      100000 NA               (163398.7 6516399, 166482.2 6513639, 155821.…
 4         64      100000 NA               (527103.2 7598541, 525136.8 7598936, 523133.…
 5         64      100000 NA               (208082.5 6807905, 210869.6 6805693, 218393.…
 6         64      100000 NA               (28219.27 6691623, 25983.92 6690898, 25870.3…
 7         64      100000 NA               (92235.33 6740841, 90324.37 6743008, 88855.2…
 8         64      100000 NA               (221028.5 6866677, 220666.4 6867071, 219770.…
 9         64      100000 NA               (753705.9 7706934, 756208.5 7709680, 758654.…
10         64      100000 NA               (101764.9 6886328, 101142.2 6890067, 102161.…
# ℹ 35 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Restriksjonsomrader_GML.gml}
}
\usage{
Naturverngrense
}
\description{
Naturverngrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
