% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Naturverngrense}
\alias{Naturverngrense}
\title{Naturverngrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 46 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 1183.26 ymin: 6472365 xmax: 1088054 ymax: 7893205
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 46 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 NA               gen             100000 (279428.2 6643125, 279462 6642494, 279579 66…
 2 NA               gen             100000 (863873 7658542, 863904.8 7658597, 864120.4 …
 3 NA               gen             100000 (410167.3 7531961, 411245.7 7531237, 412145.…
 4 NA               gen             100000 (208082.5 6807905, 210869.6 6805693, 218393.…
 5 NA               gen             100000 (28219.27 6691623, 25983.92 6690898, 25870.3…
 6 NA               gen             100000 (163398.7 6516399, 166482.2 6513639, 155821.…
 7 NA               gen             100000 (92235.33 6740841, 90324.37 6743008, 88855.2…
 8 NA               gen             100000 (221028.5 6866677, 220666.4 6867071, 219770.…
 9 NA               gen             100000 (753705.9 7706934, 756208.5 7709680, 758654.…
10 NA               gen             100000 (101764.9 6886328, 101142.2 6890067, 102161.…
# ℹ 36 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Restriksjonsomrader_GML.gml}
}
\usage{
Naturverngrense
}
\description{
Naturverngrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
