% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Naturvernomrade}
\alias{Naturvernomrade}
\title{Naturvernomrade}
\format{
\preformatted{Simple feature collection with 42 features and 4 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: 1183.26 ymin: 6472365 xmax: 1088054 ymax: 7893205
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 42 x 5
   oppdateringsdato navn    vernedato  verneform                               geometry
 * <date>           <chr>   <date>     <chr>                              <POLYGON [m]>
 1 2014-01-28       Rago    1971-01-22 NP        ((533613.6 7475526, 534740.7 7479364,~
 2 2014-01-28       Láhko   2012-12-14 NP        ((457971.3 7413120, 457970.2 7413124,~
 3 2014-01-28       Ånderd~ 2004-06-04 NP        ((585497 7677287, 583700 7679367, 583~
 4 2014-01-28       Stabbu~ 2002-12-20 NP        ((845960.9 7801580, 846043.4 7807370,~
 5 2014-01-28       Færder  2013-08-23 NP        ((243785.7 6575539, 245050.8 6577008,~
 6 2014-01-28       Sjunkh~ 2010-02-05 NP        ((494981.1 7472050, 493041.1 7472155,~
 7 2014-01-28       Foroll~ 2001-12-21 NP        ((270385.5 6942384, 271351.6 6946091,~
 8 2014-01-28       Seiland 2006-12-08 NP        ((794168.3 7831570, 793987.3 7831691,~
 9 2014-01-28       Femund~ 2003-02-21 NP        ((341250.5 6931621, 345746.7 6933494,~
10 2014-01-28       Josted~ 1991-10-25 NP        ((87209.95 6861144, 85563.37 6859917,~
# ... with 32 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Restriksjonsomrader_GML.gml}
}
\usage{
Naturvernomrade
}
\description{
Naturvernområde
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
