% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Naturvernomrade}
\alias{Naturvernomrade}
\title{Naturvernomrade}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 43 features and 4 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: 1183.26 ymin: 6472365 xmax: 1088054 ymax: 7893205
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 43 × 5
   oppdateringsdato vernedato  verneform navn                                     område
 * <date>           <date>     <chr>     <chr>                             <POLYGON [m]>
 1 2017-01-04       1971-01-22 NP        Rago           ((533613.6 7475526, 549468.5 74…
 2 2017-01-04       2012-12-14 NP        Láhko          ((457971.3 7413120, 460540.9 74…
 3 2017-01-04       2004-06-04 NP        Ånderdalen     ((585497 7677287, 584761 767333…
 4 2017-01-04       2002-12-20 NP        Stabbursdalen  ((845960.9 7801580, 846028.5 77…
 5 2017-01-04       2013-08-23 NP        Færder         ((243785.7 6575539, 243406.8 65…
 6 2017-01-04       2010-02-05 NP        Sjunkhatten    ((494981.1 7472050, 495072.8 74…
 7 2017-01-04       2001-12-21 NP        Forollhogna    ((270385.5 6942384, 270875.7 69…
 8 2017-01-04       2006-12-08 NP        Seiland        ((794168.3 7831570, 794321 7831…
 9 2017-01-04       2003-02-21 NP        Femundsmarka   ((341250.5 6931621, 341922.7 69…
10 2017-01-04       1991-10-25 NP        Jostedalsbreen ((87209.95 6861144, 84284.93 68…
# ℹ 33 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Restriksjonsomrader_GML.gml}
}
\usage{
Naturvernomrade
}
\description{
Naturvernområde
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
