% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Riksgrense}
\alias{Riksgrense}
\title{Riksgrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 40 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 245176.8 ymin: 6521821 xmax: 1121941 ymax: 7867303
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 40 × 4
   målemetode nøyaktighet oppdateringsdato                                        grense
 *      <int>       <int> <date>                                        <LINESTRING [m]>
 1         64      100000 2023-12-18       (1008977 7804953, 1011539 7802961, 1012505 7…
 2         64      100000 2023-12-18       (523677.4 7382194, 524664.7 7383241, 527543.…
 3         64      100000 2023-12-18       (580744.9 7541597, 588108.4 7547759, 592036.…
 4         64      100000 2023-12-18       (362075.8 6684849, 362007.6 6685340, 361940.…
 5         64      100000 2023-12-18       (368189.6 6712383, 366730.6 6716276, 363928.…
 6         64      100000 2023-12-18       (319002.8 6577574, 319087.6 6579087, 318962.…
 7         64      100000 2023-12-18       (353736.6 6960754, 353016.3 6966019, 352483.…
 8         64      100000 2023-12-18       (701756.5 7667791, 705043.6 7668334, 711360.…
 9         64      100000 2023-12-18       (560020.2 7491595, 563580.1 7498418, 566043.…
10         64      100000 2023-12-18       (468357.6 7221876, 468808.9 7224430, 469832.…
# ℹ 30 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000AdministrativeOmrader_GML.gml}
}
\usage{
Riksgrense
}
\description{
Riksgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
