% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Riksgrense}
\alias{Riksgrense}
\title{Riksgrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 40 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 245176.8 ymin: 6521821 xmax: 1121941 ymax: 7867303
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 40 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 2023-12-18       gen             100000 (386612.9 7096202, 394183.3 7101741, 398815 …
 2 2023-12-18       gen             100000 (449269.1 7189059, 451534.2 7193206, 453371.…
 3 2023-12-18       gen             100000 (481853 7304613, 481064.8 7308976, 480843.7 …
 4 2023-12-18       gen             100000 (412808.6 7108866, 417042.6 7107857, 424363.…
 5 2023-12-18       gen             100000 (701248.5 7613149, 705613.4 7617213, 709539.…
 6 2023-12-18       gen             100000 (560020.2 7491595, 563580.1 7498418, 566043.…
 7 2023-12-18       gen             100000 (368189.6 6712383, 366730.6 6716276, 363928.…
 8 2023-12-18       gen             100000 (319002.8 6577574, 319087.6 6579087, 318962.…
 9 2023-12-18       gen             100000 (701756.5 7667791, 705043.6 7668334, 711360.…
10 2023-12-18       gen             100000 (777865.5 7678331, 786334 7667172, 786999.9 …
# ℹ 30 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000AdministrativeOmrader_GML.gml}
}
\usage{
Riksgrense
}
\description{
Riksgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
