% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Skog}
\alias{Skog}
\title{Skog}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 429 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -52544.3 ymin: 6453900 xmax: 1114945 ymax: 7900065
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 429 × 2
   oppdateringsdato                                                               område
 * <date>                                                                  <POLYGON [m]>
 1 NA               ((140500 6999700, 141269.6 7e+06, 137003.7 7e+06, 129440.3 6995437,…
 2 NA               ((328938.2 6731425, 329890.8 6734492, 329108.9 6735220, 327166.6 67…
 3 NA               ((1027985 7805159, 1027467 7804367, 1025434 7803928, 1023721 780460…
 4 NA               ((0 6694902, 0 6703207, -2598.9 6701592, -968.18 6693999, -69.27 66…
 5 NA               ((561682 7541539, 561471.3 7541539, 559649.3 7540048, 561130.6 7538…
 6 NA               ((499224.6 7518328, 499059.5 7519561, 498735.2 7521982, 497947.6 75…
 7 NA               ((515230.7 7511197, 512230.8 7511397, 511401.4 7508991, 511230.8 75…
 8 NA               ((460664 7439736, 459600 7441500, 461000 7444700, 464900 7443600, 4…
 9 NA               ((430008.2 7414879, 431231 7411964, 431743.4 7412040, 436719 741278…
10 NA               ((429617.7 7404671, 432606 7405100, 433052.4 7405726, 433975 740702…
# ℹ 419 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
Skog
}
\description{
Skog
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
