% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{SnoIsbre}
\alias{SnoIsbre}
\title{SnoIsbre}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 25 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: 2067.32 ymin: 6675296 xmax: 771816.9 ymax: 7804397
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 25 × 2
   oppdateringsdato                                                               område
 * <date>                                                                  <POLYGON [m]>
 1 NA               ((580374.8 7541221, 581718.2 7542418, 581888.6 7542562, 580164 7545…
 2 NA               ((483173.8 7418825, 483392.3 7422451, 481773.2 7421826, 478259.1 74…
 3 NA               ((159841.7 6844159, 159201.5 6845069, 158327.9 6846312, 153656 6848…
 4 NA               ((28330.55 6695316, 29415.47 6699337, 29279.85 6702433, 29454.64 67…
 5 NA               ((124574.1 6836164, 120978.8 6838146, 119859 6835820, 118711.5 6836…
 6 NA               ((134635.5 6839884, 135823.9 6840289, 136325.5 6840599, 137074.5 68…
 7 NA               ((141194.4 6867630, 140859 6868469, 138836 6867266, 135391 6867922,…
 8 NA               ((472675.7 7321332, 470591 7322638, 470394.2 7322762, 469597.1 7323…
 9 NA               ((593836 7566609, 594055 7569180, 593687.4 7572130, 592416.3 757308…
10 NA               ((560488.3 7444281, 560568.4 7449662, 560621.5 7451394, 557845.5 74…
# ℹ 15 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
SnoIsbre
}
\description{
SnøIsbre
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
