% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Stasjon}
\alias{Stasjon}
\title{Stasjon}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 56 features and 4 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -40191.86 ymin: 6467651 xmax: 625944.6 ymax: 7594187
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 56 × 5
   oppdateringsdato målemetode nøyaktighet navn                          posisjon
 * <date>           <chr>            <int> <chr>                      <POINT [m]>
 1 NA               gen             100000 Oslo Lufthavn Gard… (283584.3 6679144)
 2 NA               gen             100000 Kongsvinger         (334071.1 6675162)
 3 NA               gen             100000 Tønsberg            (238512.8 6579249)
 4 NA               gen             100000 Larvik              (215110.7 6556333)
 5 NA               gen             100000 Koppang             (290000.8 6833010)
 6 NA               gen             100000 Hamar               (286491.4 6746722)
 7 NA               gen             100000 Dombås              (193736.8 6895785)
 8 NA               gen             100000 Kristiansand        (86015.71 6467651)
 9 NA               gen             100000 Tynset              (280967.8 6911296)
10 NA               gen             100000 Bjørnfjell          (625944.6 7594187)
# ℹ 46 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Samferdsel_GML.gml}
}
\usage{
Stasjon
}
\description{
Stasjon
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
