% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{StedsnavnTekst}
\alias{StedsnavnTekst}
\title{StedsnavnTekst}
\format{
\preformatted{Simple feature collection with 159 features and 18 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -55504.6 ymin: 6443620 xmax: 1101454 ymax: 7939718
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 159 x 19
   streng oppdateringsdato datafangstdato tekstReferanseP~ tekstReferanseP~ tegnavstand
 * <chr>  <date>           <date>                    <int>            <int>       <int>
 1 "Roms~ 2020-01-06       2019-12-10                    0                2           0
 2 "Trom~ 2020-01-06       2019-12-10                    0                2           0
 3 "Trom~ 2020-01-06       2019-12-10                    0                2           0
 4 "Stor~ 2020-01-06       2017-01-04                    0                1           0
 5 "Nord~ 2020-01-06       2017-05-15                    0                1           0
 6 "Tröö~ 2020-01-06       2019-12-10                    0                2           0
 7 "Trøn~ 2020-01-06       2019-12-10                    0                2           0
 8 "Innl~ 2020-01-06       2019-12-27                    0                0           0
 9 "Oslo~ 2017-11-10       2017-01-04                    0                0           0
10 "Ørla~ 2017-11-10       2017-01-04                    0                2           0
# ... with 149 more rows, and 13 more variables: skriftkode <int>, skrifttype <chr>,
#   referansemålestokk <int>, fulltekst <chr>, navneobjekthovedgruppe <int>,
#   navneobjektgruppe <int>, navneobjekttype <int>, språk <chr>,
#   stedsnavnsnummer <int>, språkprioritering <chr>, skrivemåtenummer <int>,
#   stedsnummer <int>, geometry <LINESTRING [m]>
}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Stedsnavn_GML.gml}
}
\usage{
StedsnavnTekst
}
\description{
StedsnavnTekst
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
