% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{StedsnavnTekst}
\alias{StedsnavnTekst}
\title{StedsnavnTekst}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 165 features and 16 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -55504.6 ymin: 6443620 xmax: 1101454 ymax: 7939718
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 165 × 17
   streng     oppdateringsdato tekstReferansePunktN…¹ tekstReferansePunktØst tegnavstand
 * <chr>      <date>                            <int>                  <int>       <int>
 1 Tromsø     2021-11-11                            0                      0           0
 2 Hammerfest 2021-11-11                            0                      0           0
 3 Honningsv… 2021-11-11                            0                      0           0
 4 Vardø      2021-11-11                            0                      0           0
 5 Kirkenes   2021-11-11                            0                      0           0
 6 Alta       2021-11-11                            0                      0           0
 7 Brønnøysu… 2021-11-11                            0                      2           0
 8 Narvik     2021-11-11                            0                      0           0
 9 Svolvær    2021-11-11                            0                      0           0
10 Mo i Rana  2021-11-11                            0                      0           0
# ℹ 155 more rows
# ℹ abbreviated name: ¹​tekstReferansePunktNord
# ℹ 12 more variables: skriftkode <int>, skrifttype <chr>, referansemålestokk <int>,
#   fulltekst <chr>, navneobjekthovedgruppe <chr>, navneobjektgruppe <chr>,
#   navneobjekttype <chr>, språk <chr>, stedsnavnnummer <int>, stedsnummer <int>,
#   skrivemåtenummer <int>, geometri <LINESTRING [m]>
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Stedsnavn_GML.gml}
}
\usage{
StedsnavnTekst
}
\description{
StedsnavnTekst
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
