% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Territorialgrense}
\alias{Territorialgrense}
\title{Territorialgrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 78 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -99551.21 ymin: 6426048 xmax: 1121941 ymax: 7962744
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 78 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 2023-12-18       gen             100000            (434193.8 7604043, 438606 7611106)
 2 2023-12-18       gen             100000          (171068.4 6489902, 152049.3 6471914)
 3 2023-12-18       gen             100000          (245176.8 6521821, 231754.6 6523881)
 4 2023-12-18       gen             100000 (927892.2 7959494, 951360.7 7962151, 952358.…
 5 2023-12-18       gen             100000 (334712.4 7369892, 348289.4 7398513, 348738.…
 6 2023-12-18       gen             100000 (316257.5 7322653, 322999 7344652, 323313.5 …
 7 2023-12-18       gen             100000 (467563.2 7654691, 489461.7 7680193, 490129.…
 8 2023-12-18       gen             100000           (63609.57 7025260, 72502.8 7030219)
 9 2023-12-18       gen             100000        (-21040.29 6465214, -21808.94 6466105)
10 2023-12-18       gen             100000 (500870.5 7689595, 527306.2 7710050, 528110.…
# ℹ 68 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000AdministrativeOmrader_GML.gml}
}
\usage{
Territorialgrense
}
\description{
Territorialgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
