% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Tettbebyggelse}
\alias{Tettbebyggelse}
\title{Tettbebyggelse}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 125 features and 4 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -51717 ymin: 6457412 xmax: 1097078 ymax: 7930522
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 125 × 5
   oppdateringsdato målemetode nøyaktighet tettbebyggelsestype           posisjon
 * <date>           <chr>            <int> <chr>                      <POINT [m]>
 1 2017-09-27       gen             100000 MindreTettbebyggel…   (820621 7676806)
 2 2017-09-27       gen             100000 MindreTettbebyggel…   (813710 7782228)
 3 2017-09-27       gen             100000 MindreTettbebyggel…   (909351 7742351)
 4 2017-09-27       gen             100000 MindreTettbebyggel…   (878538 7802690)
 5 2017-09-27       gen             100000 MindreTettbebyggel… (766364.8 7836596)
 6 2017-09-27       gen             100000 MindreTettbebyggel… (820044.4 7862737)
 7 2017-09-27       gen             100000 MindreTettbebyggel… (850567.5 7904649)
 8 2017-09-27       gen             100000 MindreTettbebyggel…   (897632 7911741)
 9 2017-09-27       gen             100000 MindreTettbebyggel… (994668.6 7840948)
10 2017-09-27       gen             100000 MindreTettbebyggel… (963129.4 7930522)
# ℹ 115 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Arealdekke_GML.gml}
}
\usage{
Tettbebyggelse
}
\description{
Tettbebyggelse
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
