% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{VegSenterlinje}
\alias{VegSenterlinje}
\title{VegSenterlinje}
\format{
\preformatted{Simple feature collection with 662 features and 8 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -58762.81 ymin: 6457040 xmax: 1097078 ymax: 7930871
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 662 x 9
   målemetode nøyaktighet medium oppdateringsdato vegkategori vegstatus vegnummer
 *      <int>       <int> <chr>  <date>           <chr>       <chr>         <int>
 1         64      100000 T      2004-01-23       E           V                39
 2         64      100000 T      2004-01-23       E           V               134
 3         64      100000 T      2004-08-01       E           V                18
 4         64      100000 T      2004-01-23       E           V                39
 5         64      100000 T      2009-11-13       E           V                18
 6         64      100000 T      2009-11-13       E           V                 6
 7         64      100000 T      2004-01-23       E           V                39
 8         64      100000 T      2004-01-23       E           V               134
 9         64      100000 T      2004-01-23       E           V               134
10         64      100000 T      2004-01-23       E           V               134
# ... with 652 more rows, and 2 more variables: motorvegtype <chr>,
#   geometry <LINESTRING [m]>
}
}
\source{
\code{Basisdata_0000_Norge_25833_N5000Samferdsel_GML.gml}
}
\usage{
VegSenterlinje
}
\description{
VegSenterlinje
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
